refactor: Estrutura o projeto dividindo o arquivo main.dart em diretorios
