+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   sqflite:
+  path: ^1.8.3
   path_provider: ^2.1.1 # Necessário para Android/iOS/Desktop
   # Adicione estas dependências para suporte a desktop e web
   sqflite_common_ffi: ^2.3.0
